High word might be std::int64_t instead of u64

+++ include/boost/int128/detail/common_mul.hpp
@@ -66,10 +66,10 @@ BOOST_INT128_FORCE_INLINE constexpr void to_words(const T& x, std::uint32_t (&wo
 
     #endif
 
-    words[0] = static_cast<std::uint32_t>(x.low & UINT32_MAX);      // LCOV_EXCL_LINE
-    words[1] = static_cast<std::uint32_t>(x.low >> 32);             // LCOV_EXCL_LINE
-    words[2] = static_cast<std::uint32_t>(x.high & UINT32_MAX);     // LCOV_EXCL_LINE
-    words[3] = static_cast<std::uint32_t>(x.high >> 32);            // LCOV_EXCL_LINE
+    words[0] = static_cast<std::uint32_t>(x.low & UINT32_MAX);                                  // LCOV_EXCL_LINE
+    words[1] = static_cast<std::uint32_t>(x.low >> 32);                                         // LCOV_EXCL_LINE
+    words[2] = static_cast<std::uint32_t>(static_cast<std::uint64_t>(x.high) & UINT32_MAX);     // LCOV_EXCL_LINE
+    words[3] = static_cast<std::uint32_t>(static_cast<std::uint64_t>(x.high) >> 32);            // LCOV_EXCL_LINE
 }
